*** empty log message ***
git-svn-id: https://pd-gem.svn.sourceforge.net/svnroot/pd-gem/trunk/Gem@100 49b2a0ca-c541-4d49-a7ef-d785e4371512

+++ src/Geos/newWave.cpp
@@ -246,7 +246,7 @@ void newWave :: setSize( int value )
     {
         case SMALL : grid = MAXGRID/4; break;
         case MEDIUM: grid = MAXGRID/2; break;
-        case LARGE : grid = MAXGRID/1.5; break;
+        case LARGE : grid = (int)(MAXGRID/1.5); break;
         case XLARGE : grid = MAXGRID; break;
     }
     if (prevGrid > grid)
